targeting/pyproject.toml
[tool.poetry]
name = "targeting"
version = "0.1.0"
description = ""
authors = ["GregShiner <greg@justifinetwork.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
opencv-python = "^4.10.0.84"
cv2-enumerate-cameras = "^1.1.16"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
